requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic>=0.42.0
requests==2.32.3
python-dotenv==1.0.1
google-genai>=1.0.0
websockets>=12.0
